Moved the dto definitions to live with their respective classes, Added a TokenCredentials class and respective DTO
