--- AllTodo.Shared/Models/DTOs/UserDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace AllTodo.Shared.Models.DTOs
-{
-    public class UserDTO
-    {
-        public int ID { get; set; }
-        public string Username { get; set; }
-        public string PhoneNumber { get; set; }
-    }
-}
