+++ AllTodo.Shared/Models/User.cs
@@ -5,6 +5,14 @@ using System.Text;
 
 namespace AllTodo.Shared.Models
 {
+
+    public class UserDTO
+    {
+        public int ID { get; set; }
+        public string Username { get; set; }
+        public string PhoneNumber { get; set; }
+    }
+
     public class User
     {
         public User(int id, Username username, PhoneNumber phone_number)
